AgenticAIStreaming/backend/requirements.txt
fastapi[standard]
gunicorn
langchain==1.2.0
langgraph==1.0.5
langchain-openai==1.1.6
python-dotenv==1.0.0
sse-starlette
httpx
firebase-admin

